chore: 合并容器和组件到一个面板
