+++ packages/gi-site/src/pages/Analysis/MetaPanel/ComponentPanel/index.less
@@ -18,7 +18,15 @@
     background-color: var(--background-color);
     margin: 6px 4px;
     padding: 4px 8px;
+    cursor: pointer;
     box-shadow: 2px 0px 4px rgba(0, 0, 0, 0.2);
+    &:hover {
+      color: var(--primary-color);
+    }
+  }
+  .gi-sub-asset-item-focusing {
+    box-shadow: 0px 0px 8px var(--primary-color);
+    color: var(--primary-color);
   }
 }
