+++ packages/gi-site/src/pages/Analysis/MetaPanel/ContainerPanel/index.less
@@ -42,7 +42,7 @@
         display: block;
       }
       .gi-render-form-assets-wrapper {
-        height: 85%;
+        height: fit-content;
         .gi-render-form-assets-input-label {
           margin-bottom: 4px;
           font-size: 13px;
@@ -54,6 +54,7 @@
         .gi-render-form-assets-input {
           width: 100%;
           height: inherit;
+          min-height: 30px;
           border-radius: 6px;
           border: 1px solid #ccc;
           cursor: text;
